fix: record dir not created if record is empty
